test(agent): classify gateway host policy as integration coverage
